prevent users unfollowing on behalf of others

+++ app/Http/Controllers/UserFollowController.php
@@ -24,6 +24,7 @@ class UserFollowController extends Controller
 
     public function destroy(User $user, Follow $follow)
     {
+        $this->authorize('delete', $follow);
         $follow->delete();
 
         return back();
